Move file out of temp utils folder
